shifted tests to proper test routine
